Refactor ConfigList
